Respect frozen layers in example models (#123)
* #119: Make sure the git example model manager caches model files

* #119: Make Model serializable

* #119: Work around compiler bug for serializable either

* #119: Fix serialization for sequential models

* #119: Fix remaining compiler errors

* #119: Implement ImmutableGraphSerializer

* #119: Update pythonString with new interfaces

* #119: Pull model from the job

* #119: Spotless

* #119: Add downloadAndConfigureExampleModel and make ExampleModel.freezeLayers more specific to each layer. Fix layer ops bug in ApplyFunctionalLayerDeltaTask.

* #119: Refactor layerGraphIsValid to use tailRecM

* #119: Document the branches of layerGraphIsValid

* #119: Don't drop the first null when loading an input layer. Reshape the input for general models if there is only one input.

* #119: Fix LoadLayersFromHDF5IntegrationTest

* #119: Cleanup after PR review

+++ example-models/src/main/kotlin/edu/wpi/axon/examplemodel/GitExampleModelManager.kt
@@ -28,12 +28,13 @@ class GitExampleModelManager : ExampleModelManager {
     /**
      * The URL to download the example models metadata from.
      */
-    var exampleModelMetadataUrl = "https://raw.githubusercontent.com/wpilibsuite/axon-example-models/master/exampleModels.json"
+    var exampleModelMetadataUrl =
+        "https://raw.githubusercontent.com/wpilibsuite/axon-example-models/master/exampleModels.json"
 
     override fun getAllExampleModels(): IO<Set<ExampleModel>> = IO {
         check(exampleModelMetadataFile.exists()) {
             "The example model metadata file (${exampleModelMetadataFile.absolutePath}) is not on " +
-                    "disk. Try updating the cache."
+                "disk. Try updating the cache."
         }
 
         ExampleModelsMetadata.deserialize(exampleModelMetadataFile.readText()).exampleModels
@@ -59,8 +60,13 @@ class GitExampleModelManager : ExampleModelManager {
         }
 
         val file = Paths.get(cacheDir.absolutePath, exampleModel.fileName).toFile()
-        file.createNewFile()
-        FileUtils.copyURLToFile(URL(exampleModel.url), file)
+
+        // Only create and download the file if it was not already in the cache
+        if (!file.exists()) {
+            file.createNewFile()
+            FileUtils.copyURLToFile(URL(exampleModel.url), file)
+        }
+
         file
     }
 }
